+++ packages/webpack-loader-runtime/README.md
@@ -1,0 +1,1 @@
+# GraphQL Tools Webpack Loader Runtime helpers
