+++ packages/webpack-loader-runtime/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@graphql-tools/webpack-loader",
+  "name": "@graphql-tools/webpack-loader-runtime",
   "version": "6.0.14",
-  "description": "A set of utils for faster development of GraphQL tools",
+  "description": "A set of utils for GraphQL Webpack Loader",
   "repository": "git@github.com:ardatan/graphql-tools.git",
   "license": "MIT",
   "sideEffects": false,
@@ -16,10 +16,6 @@
   },
   "buildOptions": {
     "input": "./src/index.ts"
-  },
-  "dependencies": {
-    "@graphql-tools/webpack-loader-runtime": "6.0.14",
-    "tslib": "~2.0.0"
   },
   "publishConfig": {
     "access": "public",
